backend/requirements_render.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
